Assignment 3.2-Sara
I added a title on the top of the page, added a text shadow effect. (h1) I added a footer with bold and italic font.  With  all of the buttons, I made a style to transform into 1.5 times the  size when you hover over them. I did changed the color of the buttons.